// src/app/components/ui/User/types/ListUsersColumnID.enum.ts
// note: value is used for sort!

export enum ListUsersColumnID {
  title = 'name',
  lastLogin = 'lastLogin',
  email = 'email',
  location = 'location',
  department = 'department',
  evaluator = 'evaluator',
  strategy = 'strategy',
  startDate = 'startDate',
  evaluationType = 'type',
  listOfContributors = 'loc',
  selfAssessment = 'sa',
  evaluationsFor = 'evaluationsFor',
  evaluationsBy = 'evaluationsBy',
  perfSummary = 'ps',
  overAllPerf = 'overAllPerf',
  action = 'action',
}